Cast net_sales to NUMERIC and tolerate the known orphan line in staging

+++ dbt/models/staging/stg_orders.sql
@@ -7,7 +7,7 @@ renamed as (
         cast(orders_id as int64)    as order_id,
         cast(customers_id as int64) as client_id,
         cast(date_date as date)     as order_date,
-        cast(net_sales as float64)  as net_sales
+        cast(net_sales as numeric)  as net_sales
     from source
 )
